modules/job-manager: move KVS restart to its own source
