+++ src/modules/job-manager/job-manager.c
@@ -17,6 +17,7 @@
 #include "job.h"
 #include "queue.h"
 #include "active.h"
+#include "restart.h"
 #include "cancel.h"
 #include "list.h"
 #include "priority.h"
@@ -111,7 +112,7 @@ static void priority_cb (flux_t *h, flux_msg_handler_t *mh,
     priority_handle_request (h, ctx->queue, msg);
 }
 
-/* active_map_f callback
+/* reload_map_f callback
  * This is called for each job encountered in the KVS during startup.
  * The 'job' struct is valid only during the callback.
  * queue_insert() increments the 'job' usecount upon successful insert.
@@ -133,7 +134,7 @@ static int restart_from_kvs (flux_t *h, struct job_manager_ctx *ctx)
 {
     int count;
 
-    if ((count = active_map (h, restart_map_cb, ctx)) < 0)
+    if ((count = restart_map (h, restart_map_cb, ctx)) < 0)
         return -1;
     flux_log (h, LOG_DEBUG, "%s: added %d jobs", __FUNCTION__, count);
     return 0;
